+++ packages/plugin-node-modules/sources/index.ts
@@ -1,26 +1,45 @@
-import {Plugin, SettingsType}      from '@yarnpkg/core';
-import {NodeModulesHoistingLimits} from '@yarnpkg/pnpify';
+import {Hooks, Plugin, SettingsType}        from '@yarnpkg/core';
+import {xfs}                                from '@yarnpkg/fslib';
+import {NodeModulesHoistingLimits}          from '@yarnpkg/pnpify';
 
-import {NodeModulesLinker}         from './NodeModulesLinker';
-import {PnpLooseLinker}            from './PnpLooseLinker';
+import {NodeModulesLinker, NodeModulesMode} from './NodeModulesLinker';
+import {getGlobalHardlinksStore}            from './NodeModulesLinker';
+import {PnpLooseLinker}                     from './PnpLooseLinker';
 
 declare module '@yarnpkg/core' {
   interface ConfigurationValueMap {
     nmHoistingLimits: NodeModulesHoistingLimits;
+    nmMode: NodeModulesMode;
   }
 }
 
-const plugin: Plugin = {
+const plugin: Plugin<Hooks> = {
+  hooks: {
+    cleanGlobalArtifacts: async configuration => {
+      const globalHardlinksDirectory = getGlobalHardlinksStore(configuration);
+      await xfs.removePromise(globalHardlinksDirectory);
+    },
+  },
   configuration: {
     nmHoistingLimits: {
-      description: `Prevent packages can be hoisted past specific levels`,
+      description: `Prevent packages to be hoisted past specific levels`,
       type: SettingsType.STRING,
       values: [
         NodeModulesHoistingLimits.WORKSPACES,
         NodeModulesHoistingLimits.DEPENDENCIES,
         NodeModulesHoistingLimits.NONE,
       ],
-      default: `none`,
+      default: NodeModulesHoistingLimits.NONE,
+    },
+    nmMode: {
+      description: `If set to "hardlinks-local" Yarn will utilize hardlinks to reduce disk space consumption inside "node_modules" directories. With "hardlinks-global" Yarn will use global content addressable storage to reduce "node_modules" size across all the projects using this option.`,
+      type: SettingsType.STRING,
+      values: [
+        NodeModulesMode.CLASSIC,
+        NodeModulesMode.HARDLINKS_LOCAL,
+        NodeModulesMode.HARDLINKS_GLOBAL,
+      ],
+      default: NodeModulesMode.CLASSIC,
     },
   },
   linkers: [